requirements/base.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-braces==1.13.0
django-cors-headers==2.2.0
django-modeltranslation==0.13b1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.8.2
fcm-client==0.2.2
idna==2.6
langdetect==1.0.7
oauthlib==2.1.0
Pillow==6.0.0
psycopg2-binary==2.7.5
pyfcm==1.4.7
PyJWT==1.6.4
python-dotenv==0.8.2
python3-openid==3.1.0
pytz==2018.3
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
six==1.12.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
fcm-client==0.2.2
tzlocal==1.5.1

transliterate==1.10.2
Unidecode==1.0.22
urllib3==1.23
toolz==0.10.0
